% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstFun.R
\name{myFirstFun}
\alias{myFirstFun}
\title{My First Function}
\usage{
myFirstFun(x)
}
\arguments{
\item{x}{A vector of quantitative data}
}
\value{
A vector of squared components
}
\description{
My First Function
}
\examples{
y <- 1:10; myFirstFun(y)
}
