% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myHyper.R
\name{myHyper}
\alias{myHyper}
\title{A Function to Mimic a Hypergeometric Distribution}
\usage{
myHyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{the number of iterations you want to run}

\item{N}{the number of total items in your population}

\item{r}{the number of success items in your population}

\item{n}{the number of items you wish to select from the population}
}
\value{

}
\description{
A Function to Mimic a Hypergeometric Distribution
}
