% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the distribution}

\item{sigma}{the standard deviation of the distribution}

\item{a}{the point for which you wish to calculate the probability below, i.e. P(Y<=a)}
}
\value{
the probability less than or equal to a, and a plot with the area shaded
}
\description{
My Normal Curve
}
