% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecefunc.R
\name{piecefunc}
\alias{piecefunc}
\title{Piecewise Function}
\usage{
piecefunc(x, coef)
}
\arguments{
\item{x}{a value for x}

\item{coef}{the coefficients of the regression items}
}
\value{
piecewise regression
}
\description{
Piecewise Function
}
